requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiomysql==0.2.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cmake==3.26.4
discord.py==2.3.0
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.3.3
future==0.18.3
idna==3.4
Jinja2==3.1.2
lit==16.0.6
llvmlite==0.40.1rc1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0rc1
numpy==1.24.3
openai==0.27.4
openai-whisper==20230314
pycparser==2.21
pydub==0.25.1
PyMySQL==1.0.3
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
soundfile==0.12.1
SpeechRecognition==3.10.0
sympy==1.12
tiktoken==0.3.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
